feat(fetch-api): proposal for support for fetch api
