+++ src/streamingHttp.browser.js
@@ -1,11 +1,12 @@
-import { isCrossOrigin, parseUrlOrigin } from './detectCrossOrigin.browser'
-import { STREAM_DATA, FAIL_EVENT, HTTP_START, STREAM_END, ABORTING, errorReport } from './events'
-import { len } from './util'
-import { parseResponseHeaders } from './parseResponseHeaders.browser'
-import { partialComplete } from './functional'
+import { fetchTransport, isFetchAvailable, streamingFetch } from './streamingHttp.fetch.browser'
+import { xhrTransport, streamingXhr } from './streamingHttp.xhr.browser'
 
 function httpTransport () {
-  return new XMLHttpRequest()
+  if (isFetchAvailable()) {
+    return fetchTransport()
+  }
+
+  return xhrTransport()
 }
 
 /**
@@ -28,121 +29,6 @@ function httpTransport () {
  * @param {boolean} withCredentials the XHR withCredentials property will be
  *    set to this value
  */
-function streamingHttp (oboeBus, xhr, method, url, data, headers, withCredentials) {
-  'use strict'
-
-  var emitStreamData = oboeBus(STREAM_DATA).emit
-  var emitFail = oboeBus(FAIL_EVENT).emit
-  var numberOfCharsAlreadyGivenToCallback = 0
-  var stillToSendStartEvent = true
-
-  // When an ABORTING message is put on the event bus abort
-  // the ajax request
-  oboeBus(ABORTING).on(function () {
-    // if we keep the onreadystatechange while aborting the XHR gives
-    // a callback like a successful call so first remove this listener
-    // by assigning null:
-    xhr.onreadystatechange = null
-
-    xhr.abort()
-  })
-
-  /**
-    * Handle input from the underlying xhr: either a state change,
-    * the progress event or the request being complete.
-    */
-  function handleProgress () {
-    if (String(xhr.status)[0] === '2') {
-      var textSoFar = xhr.responseText
-      var newText = (' ' + textSoFar.substr(numberOfCharsAlreadyGivenToCallback)).substr(1)
-
-      /* Raise the event for new text.
-
-       On older browsers, the new text is the whole response.
-       On newer/better ones, the fragment part that we got since
-       last progress. */
-
-      if (newText) {
-        emitStreamData(newText)
-      }
-
-      numberOfCharsAlreadyGivenToCallback = len(textSoFar)
-    }
-  }
-
-  if ('onprogress' in xhr) { // detect browser support for progressive delivery
-    xhr.onprogress = handleProgress
-  }
-
-  function sendStartIfNotAlready (xhr) {
-    // Internet Explorer is very unreliable as to when xhr.status etc can
-    // be read so has to be protected with try/catch and tried again on
-    // the next readyState if it fails
-    try {
-      stillToSendStartEvent && oboeBus(HTTP_START).emit(
-        xhr.status,
-        parseResponseHeaders(xhr.getAllResponseHeaders()))
-      stillToSendStartEvent = false
-    } catch (e) { /* do nothing, will try again on next readyState */ }
-  }
-
-  xhr.onreadystatechange = function () {
-    switch (xhr.readyState) {
-      case 2: // HEADERS_RECEIVED
-      case 3: // LOADING
-        return sendStartIfNotAlready(xhr)
-
-      case 4: // DONE
-        sendStartIfNotAlready(xhr) // if xhr.status hasn't been available yet, it must be NOW, huh IE?
-
-        // is this a 2xx http code?
-        var successful = String(xhr.status)[0] === '2'
-
-        if (successful) {
-          // In Chrome 29 (not 28) no onprogress is emitted when a response
-          // is complete before the onload. We need to always do handleInput
-          // in case we get the load but have not had a final progress event.
-          // This looks like a bug and may change in future but let's take
-          // the safest approach and assume we might not have received a
-          // progress event for each part of the response
-          handleProgress()
-
-          oboeBus(STREAM_END).emit()
-        } else {
-          emitFail(errorReport(
-            xhr.status,
-            xhr.responseText
-          ))
-        }
-    }
-  }
-
-  try {
-    xhr.open(method, url, true)
-
-    for (var headerName in headers) {
-      xhr.setRequestHeader(headerName, headers[headerName])
-    }
-
-    if (!isCrossOrigin(window.location, parseUrlOrigin(url))) {
-      xhr.setRequestHeader('X-Requested-With', 'XMLHttpRequest')
-    }
-
-    xhr.withCredentials = withCredentials
-
-    xhr.send(data)
-  } catch (e) {
-    // To keep a consistent interface with Node, we can't emit an event here.
-    // Node's streaming http adaptor receives the error as an asynchronous
-    // event rather than as an exception. If we emitted now, the Oboe user
-    // has had no chance to add a .fail listener so there is no way
-    // the event could be useful. For both these reasons defer the
-    // firing to the next JS frame.
-    window.setTimeout(
-      partialComplete(emitFail, errorReport(undefined, undefined, e))
-      , 0
-    )
-  }
-}
+var streamingHttp = isFetchAvailable() ? streamingFetch : streamingXhr
 
 export { httpTransport, streamingHttp }
